Move `updateScanStatus` to scb-sdk
I found that the hook-sdk as well as the parser-sdk are updating the
scans in the same way so I extracted the function `updateScanStatus`
into the scb-sdk

After moving the function there was no need to export the
`severityCount` function anymore so I removed the export for this
function
