+++ hook-sdk/nodejs/hook-wrapper.js
@@ -1,7 +1,12 @@
 const axios = require("axios");
 const { handle } = require("./hook/hook");
 const k8s = require("@kubernetes/client-node");
-const { uploadFile, severityCount, NAMESPACE, SCAN_NAME } = require("../../scb-sdk")
+const {
+  uploadFile,
+  NAMESPACE,
+  SCAN_NAME,
+  updateScanStatus
+} = require("../../scb-sdk")
 
 console.log(`Starting hook for Scan "${SCAN_NAME}"`);
 
@@ -59,41 +64,7 @@ async function updateFindings(findings) {
   await uploadFile(findingsUploadUrl, JSON.stringify(findings));
 
   // Update the scans findingStats (severities, categories, or the count) of the scan results
-  const findingCategories = new Map();
-  for (const { category } of findings) {
-    if (findingCategories.has(category)) {
-      findingCategories.set(category, findingCategories.get(category) + 1);
-    } else {
-      findingCategories.set(category, 1);
-    }
-  }
-
-  await k8sApi.patchNamespacedCustomObjectStatus(
-    "execution.securecodebox.io",
-    "v1",
-    NAMESPACE,
-    "scans",
-    SCAN_NAME,
-    {
-      status: {
-        findings: {
-          count: findings.length,
-          severities: {
-            informational: severityCount(findings, "INFORMATIONAL"),
-            low: severityCount(findings, "LOW"),
-            medium: severityCount(findings, "MEDIUM"),
-            high: severityCount(findings, "HIGH"),
-          },
-          categories: Object.fromEntries(findingCategories.entries()),
-        },
-      },
-    },
-    undefined,
-    undefined,
-    undefined,
-    { headers: { "content-type": "application/merge-patch+json" } }
-  );
-  console.log("Updated status successfully");
+  await updateScanStatus(findings);
 }
 
 async function main() {
