+++ parser-sdk/nodejs/parser-wrapper.js
@@ -2,56 +2,16 @@ const axios = require("axios");
 const { parse } = require("./parser/parser");
 const { v4: uuid } = require('uuid');
 const k8s = require("@kubernetes/client-node");
-const { severityCount, uploadFile, NAMESPACE, SCAN_NAME } = require("../../scb-sdk/nodejs/scb-sdk");
+const {
+  uploadFile,
+  NAMESPACE,
+  SCAN_NAME,
+  updateScanStatus
+} = require("../../scb-sdk/nodejs/scb-sdk");
 
 const kc = new k8s.KubeConfig();
 kc.loadFromCluster();
 const k8sApi = kc.makeApiClient(k8s.CustomObjectsApi);
-
-async function updateScanStatus(findings) {
-
-  try {
-    const findingCategories = new Map();
-    for (const { category } of findings) {
-      if (findingCategories.has(category)) {
-        findingCategories.set(category, findingCategories.get(category) + 1);
-      } else {
-        findingCategories.set(category, 1);
-      }
-    }
-
-    await k8sApi.patchNamespacedCustomObjectStatus(
-      "execution.securecodebox.io",
-      "v1",
-      NAMESPACE,
-      "scans",
-      SCAN_NAME,
-      {
-        status: {
-          findings: {
-            count: findings.length,
-            severities: {
-              informational: severityCount(findings, "INFORMATIONAL"),
-              low: severityCount(findings, "LOW"),
-              medium: severityCount(findings, "MEDIUM"),
-              high: severityCount(findings, "HIGH"),
-            },
-            categories: Object.fromEntries(findingCategories.entries()),
-          },
-        },
-      },
-      undefined,
-      undefined,
-      undefined,
-      { headers: { "content-type": "application/merge-patch+json" } }
-    );
-    console.log("Updated status successfully");
-  } catch (err) {
-    console.error("Failed to update Scan Status via the kubernetes api");
-    console.error(err);
-    process.exit(1);
-  }
-}
 
 async function extractScan() {
   try {
